moved API calls into their own files
basic refctoring, also found out here that mutate only takes one
argument to be passed into the mutateFn, so made that an object. Need to
look into why they made this choice
TODO: Also, still don't like the architecture of a lot of things. Need
to look into rethinking, or cleaning it up